Use `Py_ssize_t` instead of `int` for buffer length
This avoids a truncation warning that is treated as an error.

+++ trinity/Font/Tr2FontManager_Blue.cpp
@@ -27,7 +27,7 @@ PyObject* PyClearBuffer(PyObject *self, PyObject *args)
 		return nullptr;
 	}
     void *destPtr = buf.buf;
-    int destLen = buf.len;
+    Py_ssize_t destLen = buf.len;
 
 	if( width*4 > pitch )
 	{
